+++ backend/src/controllers/server.controller.ts
@@ -6,6 +6,7 @@ import { DbService } from "../repository/db.repository.js";
 import { success, ZodError } from "zod";
 import { createContainer, deleteContainer, getDockerStats, startContainer, stopContainer } from "../services/docker.service.js";
 import { getGameService, getManifest } from "../services/game.service.js";
+import { hasEnoughRam } from "../services/serverHelper.service.js";
 import { logger } from "../logger.js";
 import { error } from "console";
 
@@ -38,7 +39,7 @@ const buildServer = async (req:Request, res:Response) => {
         const window = Number(process.env.CREATE_SERVER_WINDOW_MINUTES ?? 60);
         const timeElapsedSinceBuild = new Date(Date.now() - window * 60 * 1000);
 
-        if (await db.countServers() <= maxTotalServers){
+        if (await db.countServers() >= maxTotalServers){
             return res.status(429).json({
                 error: "Max servers limit reached."
             });
@@ -53,6 +54,10 @@ const buildServer = async (req:Request, res:Response) => {
             return res.status(400).json("Server name already exists");
         }
 
+        const currentUsedMb = await db.sumRamAllocForActiveServers();
+        if (!hasEnoughRam(settings.core_settings.ram_alloc_mb, currentUsedMb)) {
+            return res.status(400).json({ error: "Not enough RAM available to start this server." });
+        }
 
         // Setting up the settings
         const game: GameE = settings.core_settings.game_container;
@@ -101,6 +106,11 @@ const changeServerStatus = async (req: Request, res: Response) => {
         const containerId = serverSettings.core_settings.container_id;
 
         if (action === "starting" || action === "started") {
+            const currentUsedMb = await db.sumRamAllocForActiveServers();
+            if (!hasEnoughRam(serverSettings.core_settings.ram_alloc_mb, currentUsedMb)) {
+                return res.status(400).json({ error: "Not enough RAM available to start this server." });
+            }
+
             await startContainer(containerId);
             await db.updateServerStatus(serverSettings.core_settings.server_id!, "started");
         } else if (action === "stopped" || action === "stopping"){